+++ packages/design-system/src/components/SimpleGrid/SimpleGrid.stories.tsx
@@ -27,7 +27,4 @@ SimpleGrid.args = {
   md: 'repeat(2, 1fr)',
   lg: 'repeat(3, 1fr)',
   xl: 'repeat(4, 1fr)',
-  css: {
-    width: '100%',
-  },
 }